+++ edx_proctoring/backends/tests/test_backend.py
@@ -88,7 +88,6 @@ class TestBackendProvider(ProctoringBackendProvider):
         self.last_exam = exam
         return exam.get('external_id', None) or 'externalid'
 
-    # pylint: disable=unused-argument
     def get_instructor_url(
         self, course_id, user, exam_id=None, attempt_id=None,
         show_configuration_dashboard=False, encrypted_video_review_url=None
@@ -299,7 +298,7 @@ class TestBackends(TestCase):
 
         self.assertTrue(get_attempt_mock.called)
 
-        log_format_string = ("BLOCKING ERROR: Can't find course info url for course_id=%s")
+        log_format_string = "BLOCKING ERROR: Can't find course info url for course_id=%s"
         logger_mock.assert_any_call(log_format_string, '')